refactor(atelier): split the expression nesting guard under the source-length gate
Move the byte scanners out of nesting.rs into a nesting/scan.rs submodule and
relocate the type-angle guard tests into their own integration test file so both
files that newly crossed the 350-line source gate drop back under it. Extend the
logical-operator reset test with end-to-end prefix_identifiers_in_expression
assertions so the separator and paired chains prove parser-path acceptance.
